Advanced_weather_app/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.32
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
pandas==2.2.2
